feat: Extend `Vector3d`
Give it capabilities that formerly was split between `Orientation` and
`Position` types. We decided to consolidate those into Vector3d because,
technically, all of them are just 3-vectors. The semantic
differentiation may be nice on paper but turned out to be unnecessarily
complex and all-over-the-place. Instead, have one powerful vector3d type
for all of them.

Also allow for some new, common functions like negation and substractions.

Don't add this geometricalgebra stuff and related to it, though.