chore: Relax numpy dependency to allow >1.1.19
Upon customer request who wants to use numpy to 1.1.19.

1.1.19 just incidentally, because our dependency geometricalgebra uses 1.1.19.

Upon developer request who wants to have support for numpy 2.

Effecitvely leaves us with only the intersection of numpy 1 and 2 to
play around, everything that is not supported in either is gonna break
in the other. Lucky us is that if we stick to the non-deprecated
functions, which we seem to do, from eyeballing at least, this is not an
issue with numpy specifically, apart from some potential float64 /
float32 datatype bollocks. Numpy is largely benign in this regard.